projects/chatbot/backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Backend for the chatbot project"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "crewai>=0.100.1",
    "crewai-tools>=0.36.0",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "langchain-groq>=0.2.4",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
